Split run clear-up out of the mutation supervisor
isolation.ts had grown past the 400-line target. The "is this run still
going, and delete it if not" work now lives in isolation-cleanup.ts, which
both the clear-up before a run and the list/kill/clean commands use.

Also adds a test for a folder that cannot be asked about at all: anything
other than a missing folder stops the run instead of deleting blind.
